requirements.txt
requests
python-dotenv
tabulate
